requirements.txt
python-dotenv
 praw 
 requests 
 openai 
cloudinary